refactor: rename promotion-card-styles.ts to styles.ts
