+++ src/app/main-page/promotion-card/PromotionCard.tsx
@@ -6,7 +6,7 @@ import Typography from '@material-ui/core/Typography';
 import clsx from 'clsx';
 import React from 'react';
 import { useGameLogo } from '../../../common/hooks';
-import promotionCardStyles from './promotion-card-styles';
+import promotionCardStyles from './styles';
 
 interface Props extends WithStyles<typeof promotionCardStyles> {
   content: string;
