+++ src/app/main-page/promotion-card/styles.ts
